+++ packages/backend/src/routes/users/get-following.ts
